#1329: extract maybe_rotate_epoch_v8 + publish_equal_flow_epoch_v8
Pure code-motion split of two hot-path epoch-rotation functions out
of userspace-dp/src/afxdp/types/shared_cos_lease.rs (1992 LOC) into
a directory module:

  shared_cos_lease/
    mod.rs                          1640 LOC
    rotate_epoch_v8.rs               241 LOC   (was 214-LOC body)
    publish_equal_flow_epoch_v8.rs   162 LOC   (was 142-LOC body)
    shared_cos_lease_tests.rs       1009 LOC   (relocated; bodies unchanged)

What moved

- SharedCoSQueueLease::maybe_rotate_epoch_v8(&self, now_ns: u64) —
  the per-tick rotation seqlock state machine (#1229 Phase 6 v8 +
  #1231 v5 ATOMIC-SWAP packed_granted linearization point + #1290
  v2 peer-util demand gate).
- publish_equal_flow_epoch_v8(...) — the 9-param free fn called from
  the rotation when V8RateMode::EqualFlowSuppress is active.

Both function bodies are byte-identical to master (sed -n 1497,1710p
and 1713,1854p). The only forced edits are:

- Visibility widens from inherent-private / file-private to pub(super)
  so the rotation tick path in mod.rs and the rotation->publish call
  continue to resolve across the new submodule boundary.
- Explicit #[inline] hint added on both. They each have one call site
  so the inlining decision the compiler made implicitly on master is
  expected to stay the same. #[inline] is a hint, not a codegen
  guarantee — validated by the binary-size spot check (-0.13% vs
  master, well below the 0.5% drift threshold).
- rotate_epoch_v8.rs adds an explicit
  `use super::publish_equal_flow_epoch_v8::publish_equal_flow_epoch_v8;`
  to bring the helper into lexical scope (pub(super) widens
  visibility but does not import the symbol).

Not done in this PR (out of scope, follow-ups)

- 9-param surface of publish_equal_flow_epoch_v8 NOT folded into a
  context struct — issue body's parameter-pack suggestion is a
  separate refactor.
- maybe_rotate_epoch_v8 NOT decomposed into try_claim/snapshot/publish
  phases — issue body's body-split is a state-machine rewrite, not
  pure code motion.
- god-function (>100 LOC) and >8-param cues from
  docs/engineering-style.md are NOT closed by this PR. They will be
  tracked as follow-up issues.

No algorithm change. No atomic memory ordering change. No allocation
change. Public-API re-export list in types/mod.rs is byte-identical:

  pub(super) use shared_cos_lease::{
    NOT_PARTICIPATING, PaddedVtimeSlot, SharedCoSExactBacklog,
    SharedCoSQueueLease, SharedCoSQueueVtimeFloor,
    SharedCoSRootLease, V8RateMode,
  };

Plan + adversarial plan review

- Plan doc: docs/pr/1329-shared-cos-lease-extract/plan.md
- Codex r1 PLAN-NEEDS-MINOR (4 fixes) -> r2 PLAN-NEEDS-MINOR (3 text
  fixes) -> r3 PLAN-NEEDS-MINOR (2 final text fixes) -> r4 PLAN-READY.
- AGY r1 PLAN-READY at v1; held through v2/v3/v4 text-only edits.

Validation

- cargo build --release: clean.
- cargo test --release --bin xpf-userspace-dp: 1433/1433 pass.
- 5x flake check on
  equal_flow_epoch_payload_is_visible_after_tag_under_concurrent_readers
  (a representative seqlock-rotation concurrent-reader test): 5/5 pass.
- Go suite: 32 packages ok, 0 fails (cargo test --release at workspace
  root has a pre-existing snat_contract_doc_guard failure in a
  different worktree path; unrelated to this PR).
- Binary-size spot check: 5,539,856 (this branch) vs 5,546,888 (master)
  = -7,032 bytes (-0.13%), under the 0.5% drift threshold.

Smoke deferral

Per the Wave-3 retirement-chain rule, per-PR smoke is deferred to
batch smoke at end of wave. Marker comment added after Copilot +
Codex + AGY + Claude SMR 4-of-4 attestation.

Closes #1329
