+++ docs/pr/1329-shared-cos-lease-extract/reviewer-ids.md
@@ -3,8 +3,12 @@
 Format: `<reviewer> <round> <task-id>`
 
 ## Plan review
-- codex r1 task-mpn2uvcy-65mg52
-- agy r1 adversarial-review-mpn2xtqj-jj9kge
+- codex r1 task-mpn2uvcy-65mg52  (PLAN-NEEDS-MINOR — 4 fixes)
+- agy r1 adversarial-review-mpn2xtqj-jj9kge  (PLAN-READY)
+- codex r2 task-mpn3389c-56t03t  (PLAN-NEEDS-MINOR — 3 text fixes)
+- codex r3 task-mpn385rp-meld8i  (PLAN-NEEDS-MINOR — 2 text fixes)
+- codex r4 task-mpn3i92x-6y2i2t  (PLAN-READY)
+- agy r1 PLAN-READY held from v1
 
 ## Code review
 - (pending PR open)
